Move Fastboot types to realm-server package
